1) logon hours _is_ a 32-bit not a 16-bit.  oops
2) logon to local workstation when it is a member of a domain is also
   allowed, by making domain_client_validate accept our own localservername
   as the domain name (a la MYSERVERNAME\user)

+++ source/include/rpc_samr.h
@@ -148,7 +148,7 @@ SamrTestPrivateFunctionsUser
 
 typedef struct logon_hours_info
 {
-	uint16 len; /* normally 21 bytes */
+	uint32 len; /* normally 21 bytes */
 	uint8 hours[32];
 
 } LOGON_HRS;
